Remove RestList
so this corresponded to something like [p1, p2, ..., p3 | pr]
where pr matches the rest of the list (so think car cdr)

I think this is probably possible, but it's going to involve a lot of
tedious lifetime stuff.  Which might ultimately be necessary to make
lazy match work.  However, I want to avoid massive complexity for
as long as possible.

The hope is that RestList is really only useful if you're doing
recursive list processing stuff, which should be replaced by the
"array" patterns anyway.  So maybe RestList isn't actually going
to be used in practice.