Remove todos
I don't think that * is ultimately going to be necessary.  As long as
denest exists, then a program can manually do the same thing in a much
more controlled fashion.

Also, I don't think that cow pattern is something that's going to be
easy and ultimately I don't think it actually buys much.