Remove Slot
Slot was needed for Next in the strict matcher, but with that gone now
it isn't needed.

Originally, there was a thought that pattern matching would work
differently such that there would be a need to namespace the different
things that you were matching.  What it ended up being didn't need this
functionality.